Bug 1738422 - Part 2: Update in-tree ICU to release 70.1. r=platform-i18n-reviewers,dminor
Update to ICU 70.1 by running "update-icu.sh" with "maint/maint-70" as the target.

Differential Revision: https://phabricator.services.mozilla.com/D129924

UltraBlame original commit: 6826f1884d0692195391c1bfa8154dc6a3c3c1a0

+++ intl/icu/source/common/uelement.h
@@ -56,7 +56,10 @@ typedef UBool U_CALLCONV UElementsAreEqual(const UElement e1, const UElement e2)
 
 
 
-typedef int8_t U_CALLCONV UElementComparator(UElement e1, UElement e2);
+
+
+
+typedef int32_t U_CALLCONV UElementComparator(UElement e1, UElement e2);
